Correct root certificate check.
The test that the root certificate matches the Google root should
check that the public keys match, rather than doing a bytewise
comparison of the entire certificate contents.  This allows the
certificate to be updated (which has been done once) without breaking
the comparison, as long as the public key remains the same.

Note that it's safe to check only the public key because the chain of
signatures has already been checked.