Change UserAuthType to Set of Enum
The `USER_AUTH_TYPE` tag defines a 32-bit integer bitmask with the types
of user authenticators that may be used to authorize a key [1]. That
means, the value for the tag can define a single authenticator but
also multiple authenticators.

The implementation so far only permitted values which defined exactly
one flag and would throw an `IllegalArgumentException` otherwise.

This commit changes the type of `AuthorizationList.userAuthType` from
`UserAuthType` to `Set<UserAuthType>`. That way, a value of `3` for
the `USER_AUTH_TYPE` tag evaluates to `(PASSWORD, FINGERPRINT)` instead
of raising an exception.

[1] https://source.android.com/security/keystore/tags#user_auth_type