Update root certificate check description
Contrary to what the README file states, the example does check whether the root certificate belongs to Google by comparing Google's public key (the same in all three root certificates currently in use) and the public key of the root certificate passed to the example. I've changed that part to accurately reflect what the example is doing.