Refactor status entry -> map method
Add a method to get the entry -> status map from a file instead of the
URL; and add a test for that.
Move the code that reads the status list to a new method to be used by
both the file and URL reading methods.

This way, we can test the mapping method with a local file.