Replace Google root CA cert constant with pub key.
The only checking of the root certificate that occurs is to extract the
public key and compare it to the public key of the retrieved attestation
certificate. The code can be marginally simplified by just directly
storing the root CA public key in the constants file instead.

Arguably it would be even simpler to store a raw DER byte array instead
of still keeping a PEM-encoded entry, but the PEM is easier to work with
in context of doing development or debugging with the openssl command
line.

The BEGIN PUBLIC KEY and END PUBLIC KEY statements have been stripped to
simplify direct decoding into DER.

Additionally, the error language has been updated to reflect the fact
the sample code now checks for certificate revocation.