Mitigate the certificate chain extension attack.
This changes the extraction of the attestation from the certificate chain.
Instead of unconditionally extracting the attestation in the leaf certificate
(if present), the code now walks up the certificate chain to the root, only
taking into account the last attestation extension it finds (i.e., the one
closest to the root).

This mitigates an attack in which an attacker crafts a new leaf certificate
with a seemingly good attestation and appends it to the certificate chain.