Cache status list HTTP response
The status list only needs to be fetched from the server when it becomes
stale according to the caching policy in the HTTP response. Import
OkHttp and take advantage of its RFC correct and pragmatic caching
behaviour.

The cache is stored on disk so will be reused across multiple
invocations.